Move  setupSeccompProfile and setupDaemonProces
The current implementation of setupSeccompProfile and setupDaemonProcess in
daemon_unix.go is Linux-specific. Correspondingly, move to daemon_linux.go and
implement stubs for FreeBSD.
